Desktop/copie la copie la licenta/AI_Licenta/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.3
Pillow==10.4.0
